+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - widgets
     - add `onTouch` to all container widgets
     - allow starting a traversing gesture outside a touchable widget
+    - fader: make `knobSize` dynamic
 
 ## 1.28.5